Fixed the shuffle problem
Was just a problem with the site still focusing on the shuffle button so shuffles were triggered with each enter. Changed to focus on the submit button and then added handling to stop resubmissions (from the keyboard listener and the button press). Also fixed a small bug where the settings page was forcing the site to refresh